server/requirements.txt
fastapi
uvicorn
langgraph
langchain_groq
langchain_community
python-dotenv
tavily-python
